Replicated event sourcing updates (pre 2.6.9 akka release)
