+++ journal/src/main/resources/reference.conf
@@ -4,9 +4,9 @@ akka.persistence.spanner {
   instance = "akka"
   database = "akka"
 
-  # Needs to be enabled to use the spanner plugin with active active entities, note that this will require
-  # additional table columns for replicated metadata that must be present in the schema and that a journal
-  # with replication meta enabled cannot be used for regular event sourced actors.
+  # Needs to be enabled to use the spanner plugin with replicated eventsourcing,
+  # note that this will also require additional table columns for replicated metadata
+  # that must be present in the schema
   with-meta = false
 
   journal {
